[1/2] Add in-call vibration features
Add backend implementation, implemented in UI instead of telephony
service this time.

Fix call tracking.

getActiveCall() can return null when ending calls locally before they
are answered locally, so better use the presence of an active call as
indicator of whether to do the vibration instead of using the state.